refactor(web): move date range picker to refresh components
The shared picker now lives in the supported frontend component tree, and all consumers use stable absolute imports.

Constraint: Frontend standards prohibit new shared usage from web/src/components.

Rejected: Keep the picker in the legacy tree | the generalized boundary would violate the current component migration rules

Confidence: high

Scope-risk: narrow

Directive: Keep date-range behavior changes separate from this module move.

Tested: npm test -- DateRangePicker.test.tsx --runInBand

Tested: npm run types:check

Tested: pre-commit run --files <changed frontend files>
