+++ web/src/components/dateRangeSelectors/SearchDateRangeSelector.tsx
@@ -1,4 +1,4 @@
-import { DateRangePickerValue } from "@/components/dateRangeSelectors/DateRangePicker";
+import { DateRangePickerValue } from "@/refresh-components/DateRangePicker";
 import { FiCalendar, FiChevronDown, FiXCircle } from "react-icons/fi";
 import { CustomDropdown } from "../Dropdown";
 import { timeRangeValues } from "@/app/config/timeRange";
