+++ web/src/lib/search/interfaces.ts
@@ -1,4 +1,4 @@
-import { DateRangePickerValue } from "@/components/dateRangeSelectors/DateRangePicker";
+import { DateRangePickerValue } from "@/refresh-components/DateRangePicker";
 import { Tag, ValidSources } from "../types";
 import { Agent } from "@/lib/agents/types";
