+++ web/src/lib/search/utils.ts
@@ -5,7 +5,7 @@ import {
   OnyxDocument,
   SourceMetadata,
 } from "./interfaces";
-import { DateRangePickerValue } from "@/components/dateRangeSelectors/DateRangePicker";
+import { DateRangePickerValue } from "@/refresh-components/DateRangePicker";
 
 export const buildFilters = (
   sources: SourceMetadata[],
